% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXP_Lmom.R
\name{exp_Lmom}
\alias{exp_Lmom}
\title{Fitting exponential distribution with L-moments}
\usage{
exp_Lmom(dat, threshold = NA)
}
\arguments{
\item{dat}{}

\item{threshold}{}
}
\value{
param Estimated parameters (2) and standard error returned as a list($estimate, $se).
}
\description{
Fitting exponential distribution with L-moments
}
\examples{
estimate = exp_Lmom(test_data, threshold)
threshold was added as to the function because the function sometimes set the location parameter higher than threshold causing NA values in parameter estimation
}
