% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAMMA.R
\name{gamma_bayes}
\alias{gamma_bayes}
\title{Fitting the gamma distribution with Bayesian inference}
\usage{
gamma_bayes(dat)
}
\arguments{
\item{dat}{the data that needs fitting (i.e. flood data)}
}
\value{
param Estimated parameters and standard error returned as a list($estimate, $se)
}
\description{
Function to fit the gamma distribution with bayes method
This is a dummy function because this method has not been implemented yet
}
\examples{
gamma_bayes(rgamma(1000, shape=100, scale=1))
}

