% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAMMA.R
\name{gamma_mle}
\alias{gamma_mle}
\title{Fitting the gamma distribution with MLE}
\usage{
gamma_mle(dat)
}
\arguments{
\item{dat}{the data that needs fitting (i.e. flood data)}
}
\value{
param Estimated parameters (2) and standard error returned as a list($estimate, $se)
}
\description{
Function to fit the gamma distribution with the maximum likelihood method
}
\examples{
gamma_mle(rgamma(1000, shape=100, scale=1))
}

