% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAMMA.R
\name{gamma_mom}
\alias{gamma_mom}
\title{Fitting the gamma distribution with mom}
\usage{
gamma_mom(dat)
}
\arguments{
\item{dat}{the data that needs fitting (i.e. flood data)}
}
\value{
param Estimated parameters (2) and standard error returned as a list($estimate, $se)
}
\description{
Function to fit the gamma distribution with the ordinary moments method
}
\examples{
library(FlomKartShinyApp)
estimate = gamma_mom(test_data)
FlomKartShinyApp::plot4server(test_data, param = estimate$estimate, distr = 2)
}
