% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEV.R
\name{get_posterior_gev}
\alias{get_posterior_gev}
\title{Calculating the posterior predictive distribution}
\usage{
get_posterior_gev(mmrp, mupars, spars, kpars)
}
\arguments{
\item{(mmrp, }{mupars, spars, kpars) parameters returned by gev_bayes. mupars, spars, kpars are the ensemble of param$estimate}
}
\value{
param Estimated parameters and standard error returned as a list($estimate, $se)
}
\description{
Function to calculate the posterior predictive distribution after calling gev_bayes
}
