% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUMBEL.R
\name{get_posterior_gumbel}
\alias{get_posterior_gumbel}
\title{Calculating the posterior predictive distribution}
\usage{
get_posterior_gumbel(mmrp, mupars, spars)
}
\arguments{
\item{(mmrp, }{mupars, spars) parameters returned by gumbel_bayes. mupars, spars are the ensemble of param$estimate}
}
\value{
param Estimated parameters and standard error returned as a list($estimate, $se)
}
\description{
Function to calculate the posterior predictive distribution after calling gumbel_bayes
}
\examples{
Needs example
}

