% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEARSON.R
\name{get_posterior_PEARSON}
\alias{get_posterior_PEARSON}
\title{Calculating the posterior predictive distribution}
\usage{
get_posterior_PEARSON(mmrp, mupars, spars, kpars)
}
\arguments{
\item{(mmrp, }{mupars, spars, kpars) parameters returned by pearson_bayes. mupars, spars, kpars are the ensemble of param$estimate}
}
\value{
param Estimated parameters and standard error returned as a list($estimate, $se)
}
\description{
Function to calculate the posterior predictive distribution after calling pearson_bayes
}
\examples{
Needs example
}

