% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEV.R
\name{gev_bayes}
\alias{gev_bayes}
\title{Fitting the GEV distribution with Bayesian inference}
\usage{
gev_bayes(dat)
}
\arguments{
\item{dat}{the data that needs fitting (i.e. flood data)}
}
\value{
param Estimated parameters and standard error returned as a list($estimate, $se)
}
\description{
Function to fit the GEV distribution with BayesianMCMC method
WE assume that the shape parameter only has a prior with mean zero and standard deviation 0.2 (dnorm(x[3], 0, 0.2))
}
\examples{
gev_bayes(evd::rgev(10000, loc=0, scale=1, shape=0))
}

