% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEV.R
\name{gev_mle}
\alias{gev_mle}
\title{Fitting the GEV distribution with MLE}
\usage{
gev_mle(dat)
}
\arguments{
\item{dat}{the data that needs fitting (i.e. flood data)}
}
\value{
param Estimated parameters and standard error returned as a list($estimate, $se)
}
\description{
Function to fit the GEV distribution with the maximum likelihood method
}
\examples{
gev_mle(evd::rgev(10000, loc=0, scale=1, shape=0))
}

