% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GL.R
\name{gl_Lmom}
\alias{gl_Lmom}
\title{Fitting the GL distribution with Lmom}
\usage{
gl_Lmom(dat)
}
\arguments{
\item{dat}{the data that needs fitting (i.e. flood data)}
}
\value{
param Estimated parameters and standard error returned as a list($estimate, $se)
}
\description{
Function to fit the Generalized Logistics distribution with the linear moments method
}
\examples{
library(FlomKart)
estimate = gl_Lmom(test_data)
plot_density(test_data, param = estimate, distr = "gl")
}

