% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GL.R
\name{gl_mom}
\alias{gl_mom}
\title{Fitting the GL distribution with mom}
\usage{
gl_mom(dat)
}
\arguments{
\item{dat}{the data that needs fitting (i.e. flood data)}
}
\value{
param Estimated parameters and standard error returned as a list($estimate, $se).
Standard error is not yet implemented
}
\description{
Function to fit the Generalized Logistics distribution with the ordinary moments method
}
\examples{
library(FlomKartShinyApp)
estimate = gl_mom(test_data)
FlomKartShinyApp::plot4server(test_data, param = estimate$estimate, distr = 4)
}

