% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUMBEL.R
\name{gumbel_bayes}
\alias{gumbel_bayes}
\title{Fitting the Gumbel distribution with Bayesian inference}
\usage{
gumbel_bayes(dat)
}
\arguments{
\item{dat}{the data that needs fitting (i.e. flood data)}
}
\value{
param Estimated parameters and standard error returned as a list($estimate, $se)
}
\description{
Function to fit the Gumbel distribution with BayesianMCMC method
We do not provide a prior for the BayesianMCMC function
}
\examples{
gumbel_bayes(XXX)
}

