% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUMBEL.R
\name{gumbel_Lmom}
\alias{gumbel_Lmom}
\title{Fitting the Gumbel distribution with Lmom}
\usage{
gumbel_Lmom(dat)
}
\arguments{
\item{dat}{the data that needs fitting (i.e. flood data)}
}
\value{
param Estimated parameters (2) and standard error returned as a list($estimate, $se)
}
\description{
Function to fit the Gumbel distribution with the linear moment method
}
\examples{
library(FlomKartShinyApp)
estimate = gumbel_Lmom(test_data)
FlomKartShinyApp::plot4server(test_data, param = estimate$estimate, distr = 1)
}

