% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUMBEL.R
\name{gumbel_mle}
\alias{gumbel_mle}
\title{Fitting the Gumbel distribution with MLE}
\usage{
gumbel_mle(dat)
}
\arguments{
\item{dat}{the data that needs fitting (i.e. flood data)}
}
\value{
param Estimated parameters (2) and standard error returned as a list($estimate, $se)
}
\description{
Function to fit the Gumbel distribution with the maximum likelihood method
}
\examples{
gumbel_mle(XXX)
}

