% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOGSTIC.R
\name{logist_Lmom}
\alias{logist_Lmom}
\title{Fitting the Logistic distribution with Lmom}
\usage{
logist_Lmom(dat)
}
\arguments{
\item{dat}{the data that needs fitting (i.e. flood data)}
}
\value{
param Estimated parameters and standard error returned as a list($estimate, $se)
}
\description{
Function to fit the Logistics distribution with the linear moments method
}
\examples{
library(FlomKartShinyApp)
estimate = logist_Lmom(test_data)
FlomKartShinyApp::plot4server(test_data, param = estimate$estimate, distr = 4)
}

