% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOGSTIC.R
\name{logist_mle}
\alias{logist_mle}
\title{Fitting the Logistic distribution with MLE}
\usage{
logist_mle(xdat, ydat = NULL, mul = NULL, sigl = NULL,
  mulink = identity, siglink = identity, muinit = NULL, siginit = NULL,
  show = FALSE, method = "Nelder-Mead", maxit = 10000, ...)
}
\arguments{
\item{...}{}
}
\value{
param Estimated parameters and standard error returned as a list($estimate, $se)
Standard error is not yet implemented
}
\description{
Function to fit the Logistic distribution with the maximum likelihood method.
This function was copied from Kolbjorn's initial file
}
\examples{
library(FlomKartShinyApp)
estimate = logist_mle(test_data)
FlomKartShinyApp::plot4server(test_data, param = estimate$estimate, distr = 4)
}
