% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOGSTIC.R
\name{logist_mom}
\alias{logist_mom}
\title{Fitting the Logistic distribution with mom}
\usage{
logist_mom(dat)
}
\arguments{
\item{dat}{the data that needs fitting (i.e. flood data)}
}
\value{
param Estimated parameters and standard error returned as a list($estimate, $se).
Standard error is not yet implemented
}
\description{
Function to fit the Logistics distribution with the ordinary moments method
}
\examples{
library(FlomKartShinyApp)
estimate = logistic_mom(test_data)
FlomKartShinyApp::plot4server(test_data, param = estimate$estimate, distr = 4)
}

