% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEARSON.R
\name{pearson_Lmom}
\alias{pearson_Lmom}
\title{Fitting the pearson distribution with Lmom}
\usage{
pearson_Lmom(dat)
}
\arguments{
\item{dat}{the data that needs fitting (i.e. flood data)}
}
\value{
param Estimated parameters and standard error returned as a list($estimate, $se).
Standard error is not yet implemented
}
\description{
Function to fit the pearson distribution with the linear moment method
}
\examples{
library(FlomKartShinyApp)
estimate = pearson_Lmom(test_data)
FlomKartShinyApp::plot4server(test_data, param = estimate$estimate, distr = 5)
}
