% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEARSON.R
\name{pearson_mom}
\alias{pearson_mom}
\title{Fitting the pearson distribution with mom}
\usage{
pearson_mom(dat)
}
\arguments{
\item{dat}{the data that needs fitting (i.e. flood data)}
}
\value{
param Estimated parameters and standard error returned as a list($estimate, $se)
Standard error is not yet implemented
}
\description{
Function to fit the pearson distribution with the ordinary moments method
}
\examples{
library(FlomKart)
estimate = pearson_mom(test_data)
plot_density(test_data, param = estimate, distr = "pearson")
}

