% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAMMA.R
\name{gamma_bayes}
\alias{gamma_bayes}
\title{Fitting the gamma distribution with Bayesian inference}
\usage{
gamma_bayes(dat)
}
\arguments{
\item{dat}{the data that needs fitting (i.e. flood data)}
}
\value{
param Estimated parameters and standard error returned as a list($estimate, $se)
}
\description{
Function to fit the gamma distribution with BayesianMCMC method
}
\examples{
library(FlomKartShinyApp)
estimate = gamma_bayes(test_data)
FlomKartShinyApp::plot4server(test_data, param = estimate$estimate, distr = 5)
}
