% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEV.R
\name{gev_mom}
\alias{gev_mom}
\title{Fitting the GEV distribution with mom}
\usage{
gev_mom(dat)
}
\arguments{
\item{dat}{the data that needs fitting (i.e. flood data)}
}
\value{
param Estimated parameters and standard error returned as a list($estimate, $se)
Standard error is not yet implemented
}
\description{
Function to fit the GEV distribution with the ordinary moments method
}
\examples{
library(FlomKartShinyApp)
estimate = gev_mom(test_data)
FlomKartShinyApp::plot4server(test_data, param = estimate$estimate, distr = 3)
}
\seealso{
\link{gev_Lmom}, \link{gev_mle}
}

