% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GL.R
\name{gl_bayes}
\alias{gl_bayes}
\title{Fitting the GL distribution with Bayesian inference}
\usage{
gl_bayes(dat)
}
\arguments{
\item{dat}{the data that needs fitting (i.e. flood data)}
}
\value{
param Estimated parameters and standard error returned as a list($estimate, $se)
}
\description{
Function to fit the Generalized Logistics distribution with BayesianMCMC method
We assume that the shape parameter only has a prior with mean zero and standard deviation 0.2 (dnorm(x[3], 0, 0.2))
}
\examples{
library(FlomKartShinyApp)
estimate = gl_bayes(test_data)
FlomKartShinyApp::plot4server(test_data, param = estimate$estimate, distr = 4)
}

