% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GL.R
\name{gl_mle}
\alias{gl_mle}
\title{Fitting the GL distribution with MLE}
\usage{
gl_mle(xdat, ydat = NULL, mul = NULL, sigl = NULL, shl = NULL,
  mulink = identity, siglink = identity, shlink = identity,
  muinit = NULL, siginit = NULL, shinit = NULL, show = TRUE,
  method = "Nelder-Mead", maxit = 10000, ...)
}
\arguments{
\item{dat}{the data that needs fitting (i.e. flood data)}
}
\value{
param Estimated parameters and standard error returned as a list($estimate, $se)
Standard error is not yet implemented
}
\description{
Function to fit the GL distribution with the maximum likelihood method.
This function was copied from Kolbjorn's initial file
}
\examples{
gl_mle(nsRFA::rand.genlogis(1000, 1000, 100, 0))
}

